+++ app/ui/Master/Master.tsx
@@ -12,10 +12,10 @@ import { useEffect, useState } from 'react'
 import { socket, updateSocketAuth } from '@/socket'
 import { type User } from "@supabase/supabase-js"
 import { MessageProps, Channel, ChannelHandler, ChatInputProps, ChannelUser } from '@/app/lib/types/types'
-import ChatInput from '@/app/ui/components/ChatInput'
-import PreviousMessages from '@/app/ui/components/PreviousMessages'
-import UserList from '@/app/ui/components/UserList'
-import ChannelBar from '@/app/ui/components/ChannelBar';
+import ChatInput from '@/app/ui/Chat/ChatInput'
+import PreviousMessages from '@/app/ui/Chat/PreviousMessages'
+import UserList from '@/app/ui/Users/UserList'
+import ChannelBar from '@/app/ui/Channels/ChannelBar';
 import { EmojiClickData } from 'emoji-picker-react'
 import {
     fetchAllChannelsForCurrentUser,
